API: create pipeline.py
