+++ dtoolkit/transformer/__init__.py
@@ -15,7 +15,7 @@ from dtoolkit.transformer.pandas import GetTF  # noqa
 from dtoolkit.transformer.pandas import QueryTF  # noqa
 from dtoolkit.transformer.pandas import ReplaceTF  # noqa
 from dtoolkit.transformer.pandas import SelectDtypesTF  # noqa
-from dtoolkit.transformer.sklearn import FeatureUnion  # noqa
-from dtoolkit.transformer.sklearn import make_union  # noqa
+from dtoolkit.transformer.pipeline import FeatureUnion  # noqa
+from dtoolkit.transformer.pipeline import make_union  # noqa
 from dtoolkit.transformer.sklearn import MinMaxScaler  # noqa
 from dtoolkit.transformer.sklearn import OneHotEncoder  # noqa
