+++ dtoolkit/transformer/sklearn.py
@@ -6,7 +6,6 @@ from typing import TYPE_CHECKING
 import numpy as np
 import pandas as pd
 from pandas.util._decorators import doc
-from sklearn.pipeline import FeatureUnion as SKFeatureUnion
 from sklearn.preprocessing import MinMaxScaler as SKMinMaxScaler
 from sklearn.preprocessing import OneHotEncoder as SKOneHotEncoder
 
@@ -15,7 +14,6 @@ from dtoolkit.accessor.series import cols  # noqa
 from dtoolkit.transformer._util import transform_array_to_frame
 from dtoolkit.transformer._util import transform_frame_to_series
 from dtoolkit.transformer._util import transform_series_to_frame
-from dtoolkit.transformer.base import Transformer
 
 if TYPE_CHECKING:
     from scipy.sparse import csr_matrix
@@ -23,76 +21,6 @@ if TYPE_CHECKING:
     from dtoolkit._typing import SeriesOrFrame
     from dtoolkit._typing import TwoDimArray
 
-
-class FeatureUnion(SKFeatureUnion, Transformer):
-    """
-    Concatenates results of multiple transformer objects.
-
-    See Also
-    --------
-    make_union
-        Convenience function for simplified feature union construction.
-
-    Notes
-    -----
-    Different to :obj:`sklearn.pipeline.FeatureUnion`.
-    This would let :obj:`~pandas.DataFrame` in and
-    :obj:`~pandas.DataFrame` out.
-
-    Examples
-    --------
-    >>> from dtoolkit.transformer import FeatureUnion
-    >>> from sklearn.decomposition import PCA, TruncatedSVD
-    >>> union = FeatureUnion([("pca", PCA(n_components=1)),
-    ...                       ("svd", TruncatedSVD(n_components=2))])
-    >>> X = [[0., 1., 3], [2., 2., 5]]
-    >>> union.fit_transform(X)
-    array([[ 1.5       ,  3.0...,  0.8...],
-           [-1.5       ,  5.7..., -0.4...]])
-    """
-
-    def _hstack(self, Xs):
-        if all(isinstance(i, (pd.Series, pd.DataFrame)) for i in Xs):
-            Xs = (i.reset_index(drop=True) for i in Xs)
-            return pd.concat(Xs, axis=1)
-
-        return super()._hstack(Xs)
-
-
-def make_union(
-    *transformers: list[Transformer],
-    n_jobs: int | None = None,
-    verbose: bool = False,
-) -> FeatureUnion:
-    """
-    Construct a FeatureUnion from the given transformers.
-
-    See Also
-    --------
-    FeatureUnion
-        Class for concatenating the results of multiple transformer objects.
-
-    Notes
-    -----
-    Different to :obj:`sklearn.pipeline.make_union`.
-    This would let :obj:`~pandas.DataFrame` in and
-    :obj:`~pandas.DataFrame` out.
-
-    Examples
-    --------
-    >>> from sklearn.decomposition import PCA, TruncatedSVD
-    >>> from dtoolkit.transformer import make_union
-    >>> make_union(PCA(), TruncatedSVD())
-     FeatureUnion(transformer_list=[('pca', PCA()),
-                                   ('truncatedsvd', TruncatedSVD())])
-    """
-    from sklearn.pipeline import _name_estimators
-
-    return FeatureUnion(
-        _name_estimators(transformers),
-        n_jobs=n_jobs,
-        verbose=verbose,
-    )
 
 
 class MinMaxScaler(SKMinMaxScaler):
